requirements.txt
torch
torchvision
torchaudio
diffusers
transformers
pillow
matplotlib
streamlit
tk
gradio